requirements.txt
Django==1.11.5
django-filter==1.0.4
djangorestframework==3.6.4
django-solo==1.1.2
djangorestframework-csv==2.0.0
django-logentry-admin==1.0.2
